--- src/main/kotlin/com/epam/brn/csv/UploadFromCsvStrategy.kt
@@ -1,8 +1,0 @@
-package com.epam.brn.csv
-
-import java.io.InputStream
-
-interface UploadFromCsvStrategy {
-
-    fun uploadFile(inputStream: InputStream): List<Any>
-}
